feat(vfx): 1:1 dump-accurate Force Lightning rendering (LightningEmitter parity)
Rewrite the lightning bolt renderer to mirror swkotor2.exe's LightningEmitter
(FUN_004fe770 Update / FUN_004fd050 RandomizeControlPoints / FUN_004fbed0
SpawnBranches / FUN_004f5f50 ribbon renderer / FUN_004ec080 PartEmitter::Update),
replacing the single-chord + per-frame-rerolled jitter ribbon:

- Two-level curve: coarse cubic-Bezier spline (trunc(dist*controlPTCount+0.5)+2
  control points, randomized tangents: T0 = emitter local +Z * tangentLength,
  interiors tilted +/-tangentSpread deg) re-rolled every controlPTDelay and
  cross-faded old->new when controlPTSmoothing; fine points sampled on it with
  jitter offsets re-rolled every lightningDelay. Flicker cadence is AUTHORED per
  strand, not the display frame rate (EffectBeam no longer forces a rebuild
  every frame); between rebuilds the shape rigid-follows the endpoints.
- Fine point count = trunc(dist * lightningSubDiv + 2) (engine overwrites the
  birthrate slot every frame); zigzag is a perp-rotation ANGLE step in degrees;
  lightningRadius is a no-op on the root bolt (zero base perpendicular) - both
  act on BRANCH bolts, which are now rendered: rand()%(branchCount+1) forks per
  coarse cycle with the engine's exact anchor (trunc(count*f)), short/long
  target rules, (0,0,-1) short-branch perp quirk, and f-scaled widths.
- Ribbon: per-SEGMENT camera-plane-facing quads with gapless shared edges, width
  = 2*interp(size, t=i/count) via the exact percent-stop 3-phase interpolation
  (incl. the flat-branch double-width quirk), quarter-width along-segment
  overlap extension, one full texture cell per segment, and per-segment flat
  RGBA (colorStart/Mid/End + alpha along the BOLT, engine render style 6) baked
  into a new pColor vertex attribute consumed by the LIGHTNING shader branch.
- EffectBeam: aims the beam model's +Z at the target (strand node orientations
  fan multi-strand beams at the hand), honors the EffectBeam nBodyPart node map
  (handconjure/impact/headconjure/lhand/rhand), anchors the far end at the
  target's impact node, and plays 'cast01' on attach.
- OdysseyModelNodeEmitter: controlPTSmoothing is int32 in the engine; reading it
  as float32 turned authored 1 into 1.4e-45 and disabled morph smoothing.
- GameEffect.clone() + clone-on-reapply in ModuleObject.addEffect: the engine
  copies effects on every application; re-applying one instance across an AOE
  loop re-pointed the shared beam at the last victim and the applied-guard
  swallowed repeat applications.

Harness-verified at 1280x960: v_lightnx_dur renders layered tapering streams
with forks; fine/coarse counts match the formulas (19 pts @ 5.7m subdiv 3, 3
ctrl pts); geometry holds between authored delay windows and re-rolls across
them; multi-target casts produce independent per-victim bolts. tsc 131 (0 new).

Ground truth: .planning/tmp/force-re/lightning-vfx-authored.md + Ghidra RE in
C:/Users/Alec/kotor-re/out (LightningEmitter class dumps).

+++ src/odyssey/OdysseyModelNodeEmitter.ts
@@ -66,7 +66,9 @@ export class OdysseyModelNodeEmitter extends OdysseyModelNode {
 
     //lightning
     this.branchCount = this.odysseyModel.mdlReader.readUInt32();
-    this.controlPTSmoothing = this.odysseyModel.mdlReader.readSingle();
+    // int/bool in the engine (swkotor2 emitter-def +0x60) — reading it as float32 turned the
+    // authored 1 into a denormal 1.4e-45, silently disabling coarse-spline morph smoothing.
+    this.controlPTSmoothing = this.odysseyModel.mdlReader.readUInt32();
 
     this.gridX = this.odysseyModel.mdlReader.readUInt32();
     this.gridY = this.odysseyModel.mdlReader.readUInt32();
